Unify CCHGraph and RODUACCHGraph on the shared CCHGraphBase template
The topology construction (dense node indexing, arc building with folded
via chains, TAZ member sets, inertial-flow order + CCH), path expansion
and the arc-weight convention were duplicated between the simulation and
duarouter mappers. They now live once in utils/router/CCHGraphBase<E, V>
(the GRAPH argument of CCHRouter), following the same pattern as the
router template itself.

Only the weight-fill policy stays per side: the simulation precomputes
per-arc connection permission masks for the classes in the demand and
ANDs in the live edge permissions at fill time (closure support, sparse
re-customization); duarouter walks the class successors lazily once per
class since its weights are static.
